Organize point clouds generated with cloud_node.
Organized point clouds can accelerate computations considerably. So add a new RawData::unpack() method that processes all data packets in a message and stitches them together to form an organized point cloud.

Only few changes have to be made to the previous version of RawData::unpack() to achieve this behavior. The old version of the method is temporarily kept for compatibility with transform_node. It will be removed when transform_node is adapted to output organized point clouds, too.