Transform point cloud on a per-point basis.
Until now, transform_node used the time stamp of a whole Velodyne message block to transform all points contained in this block at once.

This approach, however, is not very precise when applied to moving laser scanners. It does not consider the time stamps of the individual points, but uses the average time stamp of all points in a block.

Now, the time stamp of each point is calculated and used to transform this point from the sensor frame into the target frame.